Blog_project/requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-tinymce4-lite==1.7.4
idna==2.8
isort==4.3.17
jsmin==2.2.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
Pillow==6.0.0
pylint==2.3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
